removed functions to a helper function page
1. transformSkillRatingForDB
2. transformSkillRating
3. initializeSkillRating
